finished ft_address_print

+++ ft_low_hex_print.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_low_hex_print.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ttachi <ttachi@student.42tokyo.ja>         +#+  +:+       +#+        */
+/*   By: ttachi <ttachi@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/02 17:40:45 by ttachi            #+#    #+#             */
-/*   Updated: 2022/11/04 19:43:37 by ttachi           ###   ########.fr       */
+/*   Updated: 2022/11/04 20:12:58 by ttachi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 int	ft_low_hex_print(const char **argv, va_list ap)
 {
-	unsigned long long	x;
-	int					digit;
+	unsigned int	x;
+	int				digit;
 
 	digit = 1;
-	x = va_arg(ap, unsigned long long);
+	x = va_arg(ap, unsigned int);
 	digit = ft_hex_print(x, LOWERCASE_HEX, digit);
 	argv++;
 	return (digit);
