+++ libftprintf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libftprintf.h                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ttachi <ttachi@student.42tokyo.ja>         +#+  +:+       +#+        */
+/*   By: ttachi <ttachi@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/27 19:40:07 by ttachi            #+#    #+#             */
-/*   Updated: 2022/11/04 19:44:16 by ttachi           ###   ########.fr       */
+/*   Updated: 2022/11/04 20:12:23 by ttachi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ int		ft_putnbr_fd(int n, int fd, int digit);
 int		ft_decimal_print(const char **argv, va_list ap);
 int		ft_low_hex_print(const char **argv, va_list ap);
 int		ft_up_hex_print(const char **argv, va_list ap);
-int		ft_hex_print(unsigned long long num, char *rule, int digit);
+int		ft_hex_print(unsigned int num, char *rule, int digit);
 int		ft_address_print(const char **argv, va_list ap);
 char	*ft_strdup(const char *s1);
